+++ pkg/cmd/jira-watcher-controller/cmd.go
@@ -42,7 +42,9 @@ func NewJiraWatcherControllerCommand(name string) *cobra.Command {
 		BigQueryRefreshInterval: 1 * time.Minute,
 		// JiraSearch: "filter=105522", // CI Search - 30 Days
 		// JiraSearch: "filter=105121", // CI Search - 60 Days
-		JiraSearch: "filter=105523", // CI Search - 90 Days
+		// JiraSearch: "filter=105523", // CI Search - 90 Days
+		// TODO: 3/30/2026 - Remove in roughly 90 days (approx 7/1/2026)
+		JiraSearch: "filter=106793", // CI Search - 5 Days
 	}
 
 	ccc := controllercmd.NewControllerCommandConfig("jira-watcher-controller", version.Get(), func(ctx context.Context, controllerContext *controllercmd.ControllerContext) error {
